[refac] move controller package to top
